Improved and added tests.
Property individualNumber is now nil if a) PID length is incorrect or b) the number is not a valid integer in the range 2...999 inclusive. Else, the integer value from PID string is provided even if the PID would be invalid in other ways (date is wrong, century char is wrong, etc.).